requirements.txt
psycopg2-binary
faker
langchain>=0.1.0
langchain-community>=0.0.21
llama-cpp-python>=0.2.30